Make stdin available to non-pipe APIs
Earlier we were passing a nil input stream to APIs like `toBytes`. That
made the behavior weird for programs like "cat" which read from stdin.
Because they would just silently exit instead of reading their input
from stdin. We should either error out in such cases or allow stdin to
be read. We did the latter in this commit.